pyproject.toml
[project]
name = "langgraph-stuff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "chromadb>=1.5.5",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.3",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.3",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "tiktoken>=0.12.0",
    "uvicorn>=0.42.0",
    "wikipedia>=1.4.0",
]
